+++ android/app/src/main/kotlin/deckers/thibault/aves/AnalysisWorker.kt
@@ -70,7 +70,7 @@ class AnalysisWorker(context: Context, parameters: WorkerParameters) : Coroutine
     private fun onStart() {
         Log.i(LOG_TAG, "Start analysis worker $id")
         runBlocking {
-            FlutterUtils.initFlutterEngine(applicationContext, SHARED_PREFERENCES_KEY, CALLBACK_HANDLE_KEY) {
+            FlutterUtils.initFlutterEngine(applicationContext, SHARED_PREFERENCES_KEY, PREF_CALLBACK_HANDLE_KEY) {
                 flutterEngine = it
             }
         }
@@ -78,14 +78,15 @@ class AnalysisWorker(context: Context, parameters: WorkerParameters) : Coroutine
         try {
             initChannels(applicationContext)
 
+            val preferences = applicationContext.getSharedPreferences(SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
+            val entryIdStrings = preferences.getStringSet(PREF_ENTRY_IDS_KEY, null)
+
             runBlocking {
                 FlutterUtils.runOnUiThread {
                     backgroundChannel?.invokeMethod(
                         "start", hashMapOf(
-                            "entryIds" to inputData.getIntArray(KEY_ENTRY_IDS)?.toList(),
+                            "entryIds" to entryIdStrings?.map { Integer.parseUnsignedInt(it) }?.toList(),
                             "force" to inputData.getBoolean(KEY_FORCE, false),
-                            "progressTotal" to inputData.getInt(KEY_PROGRESS_TOTAL, 0),
-                            "progressOffset" to inputData.getInt(KEY_PROGRESS_OFFSET, 0),
                         )
                     )
                 }
@@ -179,13 +180,12 @@ class AnalysisWorker(context: Context, parameters: WorkerParameters) : Coroutine
             .setContentIntent(openAppIntent)
             .addAction(stopAction)
             .build()
-        return if (Build.VERSION.SDK_INT >= 34) {
-            // from Android 14 (API 34), foreground service type is mandatory
-            // despite the sample code omitting it at:
+        return if (Build.VERSION.SDK_INT == 34) {
+            // from Android 14 (API 34), foreground service type is mandatory for long-running workers:
             // https://developer.android.com/guide/background/persistent/how-to/long-running
-            // TODO TLAD [Android 15 (API 35)] use `FOREGROUND_SERVICE_TYPE_MEDIA_PROCESSING`
-            val type = ServiceInfo.FOREGROUND_SERVICE_TYPE_DATA_SYNC
-            ForegroundInfo(NOTIFICATION_ID, notification, type)
+            ForegroundInfo(NOTIFICATION_ID, notification, ServiceInfo.FOREGROUND_SERVICE_TYPE_DATA_SYNC)
+        } else if (Build.VERSION.SDK_INT >= 35) {
+            ForegroundInfo(NOTIFICATION_ID, notification, ServiceInfo.FOREGROUND_SERVICE_TYPE_MEDIA_PROCESSING)
         } else {
             ForegroundInfo(NOTIFICATION_ID, notification)
         }
@@ -195,14 +195,12 @@ class AnalysisWorker(context: Context, parameters: WorkerParameters) : Coroutine
         private val LOG_TAG = LogUtils.createTag<AnalysisWorker>()
         private const val BACKGROUND_CHANNEL = "deckers.thibault/aves/analysis_service_background"
         const val SHARED_PREFERENCES_KEY = "analysis_service"
-        const val CALLBACK_HANDLE_KEY = "callback_handle"
+        const val PREF_CALLBACK_HANDLE_KEY = "callback_handle"
+        const val PREF_ENTRY_IDS_KEY = "entry_ids"
 
         const val NOTIFICATION_CHANNEL = "analysis"
         const val NOTIFICATION_ID = 1
 
-        const val KEY_ENTRY_IDS = "entry_ids"
         const val KEY_FORCE = "force"
-        const val KEY_PROGRESS_TOTAL = "progress_total"
-        const val KEY_PROGRESS_OFFSET = "progress_offset"
     }
 }