+++ android/app/src/main/kotlin/deckers/thibault/aves/channel/calls/AnalysisHandler.kt
@@ -1,15 +1,14 @@
 package deckers.thibault.aves.channel.calls
 
 import android.content.Context
-import androidx.activity.ComponentActivity
 import androidx.work.ExistingWorkPolicy
-import androidx.work.OneTimeWorkRequest
 import androidx.work.OneTimeWorkRequestBuilder
 import androidx.work.WorkInfo
 import androidx.work.WorkManager
 import androidx.work.workDataOf
 import deckers.thibault.aves.AnalysisWorker
 import deckers.thibault.aves.utils.FlutterUtils
+import io.flutter.embedding.android.FlutterActivity
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import kotlinx.coroutines.CoroutineScope
@@ -19,7 +18,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 
 
-class AnalysisHandler(private val activity: ComponentActivity, private val onAnalysisCompleted: () -> Unit) : MethodChannel.MethodCallHandler {
+class AnalysisHandler(private val activity: FlutterActivity, private val onAnalysisCompleted: () -> Unit) : MethodChannel.MethodCallHandler {
     private val ioScope = CoroutineScope(SupervisorJob() + Dispatchers.IO)
 
     override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
@@ -37,10 +36,11 @@ class AnalysisHandler(private val activity: ComponentActivity, private val onAna
             return
         }
 
-        activity.getSharedPreferences(AnalysisWorker.SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
-            .edit()
-            .putLong(AnalysisWorker.CALLBACK_HANDLE_KEY, callbackHandle)
-            .apply()
+        val preferences = activity.getSharedPreferences(AnalysisWorker.SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
+        with(preferences.edit()) {
+            putLong(AnalysisWorker.PREF_CALLBACK_HANDLE_KEY, callbackHandle)
+            apply()
+        }
         result.success(true)
     }
 
@@ -53,33 +53,24 @@ class AnalysisHandler(private val activity: ComponentActivity, private val onAna
 
         // can be null or empty
         val allEntryIds = call.argument<List<Int>>("entryIds")
-        val progressTotal = allEntryIds?.size ?: 0
-        var progressOffset = 0
 
         // work `Data` cannot occupy more than 10240 bytes when serialized
-        // so we split it when we have a long list of entry IDs
-        val chunked = allEntryIds?.chunked(WORK_DATA_CHUNK_SIZE) ?: listOf(null)
-
-        fun buildRequest(entryIds: List<Int>?, progressOffset: Int): OneTimeWorkRequest {
-            val workData = workDataOf(
-                AnalysisWorker.KEY_ENTRY_IDS to entryIds?.toIntArray(),
-                AnalysisWorker.KEY_FORCE to force,
-                AnalysisWorker.KEY_PROGRESS_TOTAL to progressTotal,
-                AnalysisWorker.KEY_PROGRESS_OFFSET to progressOffset,
-            )
-            return OneTimeWorkRequestBuilder<AnalysisWorker>().apply { setInputData(workData) }.build()
+        // so we save the possibly long list of entry IDs to shared preferences
+        val preferences = activity.getSharedPreferences(AnalysisWorker.SHARED_PREFERENCES_KEY, Context.MODE_PRIVATE)
+        with(preferences.edit()) {
+            putStringSet(AnalysisWorker.PREF_ENTRY_IDS_KEY, allEntryIds?.map { it.toString() }?.toSet())
+            apply()
         }
 
-        var work = WorkManager.getInstance(activity).beginUniqueWork(
+        val workData = workDataOf(
+            AnalysisWorker.KEY_FORCE to force,
+        )
+
+        WorkManager.getInstance(activity).beginUniqueWork(
             ANALYSIS_WORK_NAME,
             ExistingWorkPolicy.KEEP,
-            buildRequest(chunked.first(), progressOffset),
-        )
-        chunked.drop(1).forEach { entryIds ->
-            progressOffset += WORK_DATA_CHUNK_SIZE
-            work = work.then(buildRequest(entryIds, progressOffset))
-        }
-        work.enqueue()
+            OneTimeWorkRequestBuilder<AnalysisWorker>().apply { setInputData(workData) }.build(),
+        ).enqueue()
 
         attachToActivity()
         result.success(null)
@@ -105,6 +96,5 @@ class AnalysisHandler(private val activity: ComponentActivity, private val onAna
     companion object {
         const val CHANNEL = "deckers.thibault/aves/analysis"
         private const val ANALYSIS_WORK_NAME = "analysis_work"
-        private const val WORK_DATA_CHUNK_SIZE = 1000
     }
 }