+++ lib/model/source/media_store_source.dart
@@ -23,6 +23,10 @@ class MediaStoreSource extends CollectionSource {
   final Set<String> _changedUris = {};
   int? _lastGeneration;
   SourceInitializationState _initState = SourceInitializationState.none;
+  bool _safeMode = false;
+
+  @override
+  set safeMode(bool enabled) => _safeMode = enabled;
 
   @override
   SourceInitializationState get initState => _initState;
@@ -46,7 +50,7 @@ class MediaStoreSource extends CollectionSource {
       analysisController: analysisController,
       directory: directory,
       loadTopEntriesFirst: loadTopEntriesFirst,
-      canAnalyze: canAnalyze,
+      canAnalyze: canAnalyze && !_safeMode,
     ));
   }
 
@@ -175,7 +179,7 @@ class MediaStoreSource extends CollectionSource {
       pendingNewEntries.clear();
     }
 
-    mediaStoreService.getEntries(knownDateByContentId, directory: directory).listen(
+    mediaStoreService.getEntries(_safeMode, knownDateByContentId, directory: directory).listen(
       (entry) {
         // when discovering modified entry with known content ID,
         // reuse known entry ID to overwrite it while preserving favourites, etc.