+++ lib/widgets/filter_grids/common/filter_grid_page.dart
@@ -191,12 +191,10 @@ class _FilterGrid<T extends CollectionFilter> extends StatefulWidget {
 
 class _FilterGridState<T extends CollectionFilter> extends State<_FilterGrid<T>> {
   TileExtentController? _tileExtentController;
-  final DoubleBackPopHandler _doubleBackPopHandler = DoubleBackPopHandler();
 
   @override
   void dispose() {
     _tileExtentController?.dispose();
-    _doubleBackPopHandler.dispose();
     super.dispose();
   }
 
@@ -212,16 +210,12 @@ class _FilterGridState<T extends CollectionFilter> extends State<_FilterGrid<T>>
     );
     return AvesPopScope(
       handlers: [
-        (context) {
-          final selection = context.read<Selection<FilterGridItem<T>>>();
-          if (selection.isSelecting) {
-            selection.browse();
-            return false;
-          }
-          return true;
-        },
-        TvNavigationPopHandler.pop,
-        _doubleBackPopHandler.pop,
+        APopHandler(
+          canPop: (context) => context.select<Selection<FilterGridItem<T>>, bool>((v) => !v.isSelecting),
+          onPopBlocked: (context) => context.read<Selection<FilterGridItem<T>>>().browse(),
+        ),
+        tvNavigationPopHandler,
+        doubleBackPopHandler,
       ],
       child: TileExtentControllerProvider(
         controller: _tileExtentController!,