+++ plugins/aves_model/lib/src/actions/explorer.dart
@@ -1,0 +1,4 @@
+enum ExplorerAction {
+  addShortcut,
+  setHome,
+}