+++ test_driver/driver_shaders.dart
@@ -26,8 +26,7 @@ Future<void> configureAndLaunch() async {
     ..enableBlurEffect = true
     // navigation
     ..keepScreenOn = KeepScreenOn.always
-    ..homePage = HomePageSetting.collection
-    ..homeCustomCollection = {}
+    ..setHome(HomePageSetting.collection)
     ..enableBottomNavigationBar = true
     // collection
     ..collectionSectionFactor = EntryGroupFactor.album